less warnings
git-svn-id: http://svn.miranda-ng.org/main/trunk@11224 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/AddContactPlus/src/addcontact.cpp
@@ -126,7 +126,7 @@ bool AddContactDlgAccounts(HWND hdlg, AddDialogParam *acs)
 			continue;
 
 		cbei.pszText = pAccounts[i]->tszAccountName;
-		GetTextExtentPoint32(hdc, cbei.pszText, mir_tstrlen(cbei.pszText), &textSize);
+		GetTextExtentPoint32(hdc, cbei.pszText, (int)mir_tstrlen(cbei.pszText), &textSize);
 		if (textSize.cx > cbWidth) cbWidth = textSize.cx;
 		HICON hIcon = (HICON)CallProtoService(pAccounts[i]->szModuleName, PS_LOADICON, PLI_PROTOCOL | PLIF_SMALL, 0);
 		cbei.iImage = cbei.iSelectedImage = ImageList_AddIcon(hIml, hIcon);
