+++ plugins/TooltipNotify/src/Tooltip.cpp
@@ -147,7 +147,7 @@ void CTooltip::Validate()
 	SIZE Size;
 	HDC hDC = GetDC(m_hWnd);
 	SelectObject(hDC, m_hFont);
-	GetTextExtentPoint32(hDC, m_szText, mir_tstrlen(m_szText), &Size);
+	GetTextExtentPoint32(hDC, m_szText, (int)mir_tstrlen(m_szText), &Size);
 	SetWindowPos(m_hWnd, 0, 0, 0, Size.cx+6, Size.cy+4, 
 		SWP_NOZORDER | SWP_NOMOVE | SWP_NOACTIVATE | SWP_NOREDRAW);
 	ReleaseDC(m_hWnd, hDC);
